+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -114,6 +114,11 @@ New checks
 
   FIXME: add release notes.
 
+- New :doc:`hooks-account-conv-pure
+  <clang-tidy/checks/hooks-account-conv-pure>` check.
+
+  FIXME: add release notes.
+
 - New :doc:`hooks-entry-point-recursion
   <clang-tidy/checks/hooks-entry-point-recursion>` check.